use different class based on scala version
