+++ kernel/src/main/scala/cats/kernel/Enumerable.scala
@@ -1,7 +1,6 @@
 package cats
 package kernel
 
-import scala.collection.immutable.Stream
 import scala.{specialized => sp}
 
 /**
@@ -39,34 +38,6 @@ trait Previous[@sp A] extends PartialPrevious[A] {
   def partialOrder: PartialOrder[A]
   def previous(a: A): A
   override def partialPrevious(a: A): Option[A] = Some(previous(a))
-}
-
-trait PartialPreviousUpperBounded[@sp A] extends PartialPrevious[A] with PartialNext[A] with UpperBounded[A] {
-
-  /** Enumerate the members in descending order. */
-  def membersDescending: Stream[A] = {
-    def loop(a: A): Stream[A] =
-      partialPrevious(a) match {
-        case Some(aa) => aa #:: loop(aa)
-        case _        => Stream.empty
-      }
-    maxBound #:: loop(maxBound)
-  }
-
-}
-
-trait PartialNextLowerBounded[@sp A] extends PartialPrevious[A] with PartialNext[A] with LowerBounded[A] {
-
-  /** Enumerate the members in ascending order. */
-  def membersAscending: Stream[A] = {
-    def loop(a: A): Stream[A] =
-      partialNext(a) match {
-        case Some(aa) => aa #:: loop(aa)
-        case _        => Stream.empty
-      }
-    minBound #:: loop(minBound)
-  }
-
 }
 
 /**
